test(ml): #838 stage 테스트 구조와 fixture 재사용성 개선
