+++ .agent/specs/838.md
@@ -1,0 +1,73 @@
+# ML stage 테스트 구조와 fixture 재사용성 개선
+
+## Goal
+
+ML stage 테스트를 행위 단위 파일로 분리하고 공통 artifact fixture helper를 재사용하게 만들어, 실패 원인과 fixture 영향 범위를 더 빠르게 파악할 수 있게 한다.
+
+## Background
+
+Issue #838은 1,000줄 이상으로 커진 ML stage 테스트 파일과 반복되는 artifact JSON 생성 코드, optional Airflow DAG import smoke 테스트의 책임 경계를 문제로 지적한다.
+
+이슈가 지목한 변경 전 대상 파일:
+
+- `ml/tests/stages/test_flow_splitting_main.py`
+- `ml/tests/stages/test_draft_generation_main.py`
+- `ml/tests/stages/test_evaluation_main.py`
+- `ml/tests/stages/test_publish_candidate_main.py`
+- `ml/tests/stages/test_draft_generation_workflow_graph.py`
+- `ml/tests/test_pipeline.py`
+
+분리 후 현재 대상 파일:
+
+- `ml/tests/stages/test_flow_splitting_stage_run.py`
+- `ml/tests/stages/test_flow_splitting_confidence.py`
+- `ml/tests/stages/test_flow_splitting_grouping.py`
+- `ml/tests/stages/test_flow_splitting_labels.py`
+- `ml/tests/stages/test_flow_splitting_label_quality.py`
+- `ml/tests/stages/test_draft_generation_intents_io.py`
+- `ml/tests/stages/test_draft_generation_run.py`
+- `ml/tests/stages/test_draft_generation_candidate_metrics.py`
+- `ml/tests/stages/test_draft_generation_workflows.py`
+- `ml/tests/stages/test_draft_generation_slots.py`
+- `ml/tests/stages/test_evaluation_stage_run.py`
+- `ml/tests/stages/test_evaluation_benchmarks.py`
+- `ml/tests/stages/test_evaluation_quality_gates.py`
+- `ml/tests/helpers/artifacts.py`
+- `ml/tests/helpers/draft_generation.py`
+- `ml/tests/helpers/evaluation.py`
+- `ml/tests/test_domain_pack_generation_dag.py`
+
+## Scope
+
+- 1,000줄 이상 ML stage 테스트를 입력/출력, edge case, artifact IO, metric/quality gate 등 의미 있는 하위 파일로 분리한다.
+- stage 테스트에서 반복되는 manifest, JSON artifact, cluster/preprocessed fixture 작성 helper를 `ml/tests` 하위 공통 helper로 모은다.
+- Airflow DAG import 테스트는 CI에서 실행되는 import 책임과 Airflow 미설치 환경의 optional smoke 책임이 테스트 이름과 skip reason에서 구분되게 한다.
+- 기존 테스트의 검증 대상과 assertion은 유지한다.
+
+## Non-Goals
+
+- ML pipeline stage production logic 변경은 하지 않는다.
+- artifact schema, DAG stage 순서, Airflow runtime behavior를 바꾸지 않는다.
+- coverage threshold나 CI workflow 정의를 변경하지 않는다.
+
+## Affected Area
+
+- ML tests: `ml/tests/stages/`, `ml/tests/test_pipeline.py`
+- ML test helpers: `ml/tests/helpers/`
+- ML pipeline stage order는 기존 `ingestion -> preprocessing -> representation -> intent_discovery -> flow_splitting -> draft_generation -> evaluation -> publish_candidate`를 유지한다.
+
+## Requirements
+
+1. `test_flow_splitting_main.py`, `test_draft_generation_main.py`, `test_evaluation_main.py`는 1,000줄 미만의 행위 단위 test files로 분리되어야 한다.
+2. 공통 artifact fixture 작성은 새 helper에서 제공하고, 최소 두 개 이상의 stage test module이 이를 사용해야 한다.
+3. Airflow DAG import smoke 테스트는 Airflow 미설치 skip 조건과 대체 import 검증 경로를 명시해야 한다.
+4. 변경 후 `uv run pytest` 또는 관련 좁은 `uv run pytest ...` 명령이 통과해야 한다.
+
+## Validation
+
+- `cd ml && uv run pytest tests/test_pipeline.py tests/stages`
+- 필요 시 `cd ml && uv run ruff check tests`
+
+## Open Questions
+
+- 없음. 이슈에 별도 댓글이나 추가 요구사항은 없었다.
